add component default value
add component default value

+++ RenderBird/core/src/typeinfo.cpp
@@ -1,13 +1,15 @@
 #include "typeinfo.h"
 
+
+
 void GlobalTypeList::RegisterTypeInfo(TypeInfo* typeInfo)
 {
-	int typeId = typeInfo->GetTypeId();
+	size_t typeId = typeInfo->GetTypeId();
 	assert(m_typeMap.find(typeId) == m_typeMap.end());
 	m_typeMap[typeId] = typeInfo;
 }
 
-TypeInfo* GlobalTypeList::GetTypeInfo(int typeId)
+TypeInfo* GlobalTypeList::GetTypeInfo(size_t typeId)
 {
 	TypeInfoMap::iterator itr = m_typeMap.find(typeId);
 	if (itr != m_typeMap.end())
